moved Note from refactor to refactor.util
